// vietnam/assets/ScriptCore/protocol/player.d.ts
export namespace player {
// ...
    interface IPlayerData {
// ...
        Coin?: (number|Long|null);
// ...
        Diamond?: (number|Long|null);
// ...
    }
// ...
    interface IItemInfo {

        /** ItemInfo ItemId */
        ItemId?: (number|null);
// ...
    }
// ...
}